requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0
pyarrow>=12.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Technical Analysis
ta-lib>=0.4.25
pandas-ta>=0.3.14b

# Backtesting
backtrader>=1.9.78.123
bt>=0.2.9

# State Machine
python-statemachine>=2.1.0

# GUI
streamlit>=1.28.0
plotly>=5.15.0
dash>=2.14.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
click>=8.1.0

# File Processing
pathlib2>=2.3.7

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
